+++ src/core.js
@@ -2,11 +2,11 @@ var m4qVersion = "v@@VERSION. Built at @@TIME";
 var regexpSingleTag = /^<([a-z][^\/\0>:\x20\t\r\n\f]*)[\x20\t\r\n\f]*\/?>(?:<\/\1>|)$/i;
 
 var matches = Element.prototype.matches
-    || Element.prototype["matchesSelector"]
-    || Element.prototype["webkitMatchesSelector"]
-    || Element.prototype["mozMatchesSelector"]
-    || Element.prototype["msMatchesSelector"]
-    || Element.prototype["oMatchesSelector"];
+    || Element.prototype.matchesSelector
+    || Element.prototype.webkitMatchesSelector
+    || Element.prototype.mozMatchesSelector
+    || Element.prototype.msMatchesSelector
+    || Element.prototype.oMatchesSelector;
 
 var $ = function(selector, context){
     return new $.init(selector, context);
@@ -44,7 +44,7 @@ $.extend = $.fn.extend = function(){
     for ( ; i < length; i++ ) {
         if ( ( options = arguments[ i ] ) != null ) {
             for ( name in options ) {
-                if (options.hasOwnProperty(name) && !not(options[name])) target[ name ] = options[ name ];
+                if (options.hasOwnProperty(name)) target[ name ] = options[ name ];
             }
         }
     }
@@ -52,4 +52,30 @@ $.extend = $.fn.extend = function(){
     return target;
 };
 
-if (typeof window["hideM4QVersion"] === "undefined") console.info("m4q "+$.version);
+$.assign = function(){
+    var options, name,
+        target = arguments[ 0 ] || {},
+        i = 1,
+        length = arguments.length;
+
+    if ( typeof target !== "object" && typeof target !== "function" ) {
+        target = {};
+    }
+
+    if ( i === length ) {
+        target = this;
+        i--;
+    }
+
+    for ( ; i < length; i++ ) {
+        if ( ( options = arguments[ i ] ) != null ) {
+            for ( name in options ) {
+                if (options.hasOwnProperty(name) && options[name] !== undefined) target[ name ] = options[ name ];
+            }
+        }
+    }
+
+    return target;
+};
+
+if (typeof window.hideM4QVersion === "undefined") console.info("m4q "+$.version);