+++ src/mode.js
@@ -1,1 +1,2 @@
+/* jshint -W097 */
 'use strict';