+++ src/promise.js
@@ -187,7 +187,7 @@
         if (typeof resolver !== 'function')
             throw new TypeError('Promise constructor takes a function argument');
 
-        if (!this instanceof Promise)
+        if (!(this instanceof Promise))
             throw new TypeError('Failed to construct \'Promise\': Please use the \'new\' operator, this object constructor cannot be called as a function.');
 
         this.then_ = [];
@@ -309,7 +309,7 @@
         });
     };
 
-    if (typeof  global['Promise'] === "undefined") {
+    if (typeof  global.Promise === "undefined") {
         global.Promise = Promise;
     }
 }(window));