+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.0.6
 + [x] Init: fix minor bug for creating elements in context
 + [x] Ajax: add parameter `contentType`. If this param has value `false`, `Content-type` can't be defined.
++ [x] Animation: a new engine for animation. The `animate` function changed. Now function receives one argument.
 
 ### 1.0.5
 + [x] Events: fix firing events, Metro 4 issue 1476